interim save: separate out the provider code a bit
